refactored iva check to test service
